using doubles over floats

+++ src/log_regression-Brian/log_regression.c
@@ -8,7 +8,7 @@ typedef struct Term {
 // a coefficient and exponent for predicting future y values given x values.
 // The given "function" is literally the two variables, use them with this formula:
 // y = ln(x*exponent)*coefficient
-Term log_regression(int n, float *x, float *y) {
+Term log_regression(int n, double *x, double *y) {
     double sumx=0; // sum of all floats in x.
     double sumy=0; // log applied to each value in the y array, then sum of that.
     double sumxy=0; // first multiplies each value in x with the log of each value in y, and sums that.
@@ -17,7 +17,7 @@ Term log_regression(int n, float *x, float *y) {
     // "Your ancestors called it Magic, but you call it Math.
     // I come from a land where they are one and the same." -  Thor Odinson
     for(int i=0; i<=n-1; i++) {
-        float Y = log(y[i]);
+        double Y = log(y[i]);
         sumx  += x[i];
         sumx2 += x[i]*x[i];
         sumy  += Y;
